mcp-semantic-search/pyproject.toml
[project]
name = "roam-semantic-mcp"
version = "0.1.0"
description = "MCP server exposing Roam Semantic Search querying tools"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]",
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
]

[project.scripts]
roam-semantic-mcp = "server:main"

[tool.uv]
dev-dependencies = []

